Implemented the CAE.
The ContractingAutoencoder class subclasses the Autoencoder and is
used the same way, except for its cost, that need to be added the
result of a call to contraction_penalty().

Also added example_cae.py that shows how to build one, as well as
a cae/dae stack (see framework.autoencoder.build_denoising_stack).