Add sparse TERRY experiment script.
This is work in progress; I'm just pushing this now for the sake of discussions
in class later today.

Basically, this is a fully functional sparse experiment script if you change
'terry' to 'ule', which discards features which are absent in any of the three
data subsets, applies a PCA (via a sparse implementation), runs it through a
denoising autoencoder, applies another (nonsparse, assuming dimensionality is
greatly reduced at this point) PCA, and creates a submission.

I believe the only outstanding issue here is that I can't seem to find a way
to subtract the feature means from a sparse (n,d) matrix without constructing
any intermediary dense (n,d) matrices.  Currently, this sets the baseline
memory requirement at around 8.64 GB, which I guess isn't that unreasonable...