Script for TERRY PCA model selection based on ALC.
For each possible subset of the set of classes, we train a PCA on 'train' examples
which are labeled as belonging to one of those classes; we then compute the ALC on
a few random subsets of this data.  This process is repeated for different values of
a hyperparameter -- in this case, the number of components in the PCA.  The results
are stored in a file, which we can later use for a sort of layer-wise model selection,
and then repeat the process on a second layer (e.g. a DAE).